Refactor InputWithLabel component into its own file
