+++ src/InputWithLabel.tsx
@@ -1,0 +1,67 @@
+import * as React from 'react';
+import styled from 'styled-components';
+
+const StyledLabel = styled.label`
+  border-top: 1px solid #171212;
+  border-left: 1px solid #171212;
+  padding-left: 5px;
+  font-size: 24px;
+`;
+
+const StyledInput = styled.input`
+  border: none;
+  border-botttom: 1px solid #171212;
+  background-color: transparent;
+
+  font-size: 24px;
+`;
+
+type InputWithLabelProps = {
+  id: string;
+  value: string;
+  type?: string;
+  onInputChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  isFocused?: boolean;
+  children: React.ReactNode;
+};
+
+const InputWithLabel: React.FC<InputWithLabelProps> = ({
+  id,
+  value,
+  type = 'text',
+  onInputChange,
+  isFocused,
+  children,
+}) => {
+  // Create a ref with React's useRef Hook
+  const inputRef = React.useRef<HTMLInputElement>(null);
+
+  // Opt into React's lifecycle with React's useEffect Hook
+  React.useEffect(() => {
+    if (isFocused && inputRef.current) {
+      // Execute its focus programmatically as a side-effect, but only if isFocusedis set and the current property is existent
+      inputRef.current.focus();
+    }
+  }, [isFocused]);
+
+  return (
+    <>
+      <StyledLabel htmlFor={id}>{children}</StyledLabel>
+      &nbsp;
+      {/* Pass ref to element's JSX-reserved ref attribute */}
+      <StyledInput
+        ref={inputRef}
+        id={id}
+        type={type}
+        value={value}
+        // autoFocus={isFocused}
+        onChange={onInputChange}
+        // className={styles.input}
+      />
+    </>
+  );
+};
+// note that `autoFocus` is a shorthand for `autoFocus={true}`
+// every attribute that is set to `true` can use this shorthand
+
+export default InputWithLabel;
